pyproject.toml
[project]
name = "day5-passgen"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.1.0",
    "gunicorn>=23.0.0",
    "python-dotenv>=1.1.0",
]
